requirements.txt
vectorbt==0.28.1
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv
